Working project with color toggle

+++ src/time.c
@@ -6,7 +6,7 @@
 /*   By: jbagger <jbagger@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/04/20 14:36:37 by jbagger           #+#    #+#             */
-/*   Updated: 2023/04/24 07:16:24 by jbagger          ###   ########.fr       */
+/*   Updated: 2023/05/04 16:46:44 by jbagger          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 int	start_time(t_data *data)
 {
 	struct timeval	time;
-	
+
 	if (gettimeofday(&time, NULL))
 	{
 		error("gettimeofday() failed\n", data, 8);
@@ -25,15 +25,15 @@ int	start_time(t_data *data)
 	return (0);
 }
 
-long long	time_since(long long start)
+long int	time_since(long int start)
 {
 	return (time_now() - start);
 }
 
-long long	time_now(void)
+long int	time_now(void)
 {
 	struct timeval	time;
-	
+
 	gettimeofday(&time, NULL);
 	return (time.tv_sec * 1000 + time.tv_usec / 1000);
 }
